+++ components/GridToolBox/index.js
@@ -11,7 +11,6 @@ import DatePickerBetween from './../DatePicker/Between';
 import DateTimePickerBetween from '../DateTimePicker/Between';
 import ConfirmDialog from '../ConfirmDialog';
 import StandardDialog from '../Popup';
-import StandardButton from '../StandardButton';
 import { Button } from 'reactstrap';
 import { formatIso } from 'material-ui/DatePicker/dateUtils';
 import { formatTime } from 'material-ui/TimePicker/timeUtils';
@@ -355,8 +354,14 @@ class GridToolBox extends Component {
 
     renderAdvancedButton() {
         let tooltipContent = this.getTooltip();
-        let el = <div key={Math.random()} className={classnames(style.toolbarElement, style.tableCell, style.advancedSearchIconWrapper)}>
-            <button className={classnames(style.toolbarElement, style.noRightMargin, style.advancedSearchIcon)} onClick={this.toggleAdvancedSearch}>&nbsp;</button>
+        let el = <div key='toggleAdv' className={classnames(style.toolbarElement, style.tableCell, style.advancedSearchIconWrapper)}>
+            <div className={classnames(style.noRightMargin, style.advancedSearchIcon)} onClick={this.toggleAdvancedSearch}>
+                <div className={style.barWrap}>
+                    <div className={style.bar} />
+                    <div className={style.bar} />
+                    <div className={style.bar} />
+                </div>
+            </div>
             {tooltipContent.length ? <div className={style.advancedSearchPopOver}>
                 {tooltipContent}
             </div> : null}
@@ -428,6 +433,10 @@ class GridToolBox extends Component {
             leftSide = hasSelectedOrChecked ? <span className={style.link}>Show buttons</span> : 'Filter by';
         }
 
+        let showSearchBtn = (this.props.filterElements.find(f => {
+            return f.type === filterElementTypes.searchBtn;
+        }) !== undefined) || (!this.state.showFiltersPopup && !this.props.filterAutoFetch && Object.keys(this.state.filters).length > 0);
+
         return (
             <div className={classnames(style.toolbarWrap, style.table, style.fixedHeight)}>
                 <div className={classnames(style.toolbarElement, style.label, labelClass, style.tableCell)} onClick={toggle}>
@@ -450,9 +459,9 @@ class GridToolBox extends Component {
                             }
                         })}
                         {this.renderAdvanced()}
-                        {!this.state.showFiltersPopup && !this.props.filterAutoFetch && Object.keys(this.state.filters).length > 0 &&
+                        {showSearchBtn &&
                             <div key='searchBtn' className={classnames(style.toolbarElement, style.tableCell)}>
-                                <StandardButton onClick={this.applyFilters} styleType='secondaryDark' label='Apply Search' />
+                                <div onClick={this.applyFilters} className={style.searchIcon} />
                             </div>}
                         {this.state.hasActiveFilters &&
                             <div className={classnames(style.toolbarElement, style.tableCell)}>
@@ -690,7 +699,8 @@ GridToolBox.propTypes = {
                 filterElementTypes.datePickerBetween,
                 filterElementTypes.dateTimePickerBetween,
                 filterElementTypes.customSearch,
-                filterElementTypes.clear
+                filterElementTypes.clear,
+                filterElementTypes.searchBtn
             ]).isRequired,
             // Common
             placeholder: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),